Implement break, continue and let control flow statements bubble up though blocks
This involved refactoring a lot of the executor code, in order to avoid boilerplate I am using a Result to represent either an Error or a Control Flow statement that means the current block should exit early. This means I can use the ? syntax to return out of the function early, and allow the control flow statements to bubble up to points in the code where they should be handled. A bit confusing as Some `Err`s are not really errors, but I think its worth the added readability.

I settled on the name Exception to mean either an error or early return from an expression. This really means Error in some languages, but I couldn't think of a better one and currently in rust that name is unclaimed.