Rethink on commands and operators, now the exec statement only returns stdout of the command
Stderr is always redirected to the parent process output, currently there is no way to change this. Piping a command will forward its stdout to the next command, stderr will still all be forwarded to the parent output.

Exit code handling is unchanged.

This makes the command interface simpler, and in an ideal world the stderr is just use output that the script can safely ignore so this is a good default I think. Will need to add features to pipe or save the stderr of a command instead of just forwarding it to the output, as this is a reasonable thing to want a bit more control on - especially as we often cannot control the behavior of the commands we use to do exactly what we want.

This also has removed the ability to pipe output to a file - it was a bit ugly and was complicating the code plus I didn't like how a string kinda just represented a file by convention. Will need to come up with something here as well.