Implement capture options for running commands
This means that when you exec a command pipeline, you can now add options in square brackets next to the command in order to capture stderr / exit code and also suppress that stuff from getting forwarded to the process output.

This was quite a big refactor as we now do not consider something like `command1` => `command2` a valid expression anymore. Individual commands are fine but piping them together only works in an execute statement.

Also there is a new FileHandle type, not super happy about it but as Command could kinda represent Files previously its not really adding much - but just the type is more obvious now.