Restructure code - move more things into components to avoid lots of global exports
Quite a lot of the components needed to be used by other things and were exported to the whole crate. I decided to move a few things inside the components folder, and nest all the existing stuff one level deeper.

I think this makes it easier to know what interacts with the components directly and what doesn't.