Change executor to trait to avoid dynamic boxed values
I think this is more idiomatic rust, instead of relying on a vtable and passing around a trait as a value. Took me a bit to figure out this is what associated traits are for :facepalm:.

The integration tests are a little more complication now, mainly because I think just mocking out the Write and Read methods is way less useful tha using a simple concrete type for them.

Mockall doesn't support partially mocking a type. So my solution was to create my own mock implementation of the executor, which would then pass along any calls to methods I really wanted mocked to another struct that was mocked by mockall - but only contained the methods I wanted to mock.