Make any variables created other than `var mut` not mutable.
This means loop variable, function args and captured variables are not mutable and there is no way to make them mutable. I didn't want to add the complexity to allow them to be mutable, and it seems like it wouldn't really provide much benefit.

Is someone really did want to mutate one of these variables (which in general I would recommend not doing), it would be quite trivial to declare a new variable with the same name anyway.