Add post process step that catches type errors earlier
This step runs after parsing but before execution, it potentially could be part of the parsing step but having access to the entire parsed expression tree could be useful here maybe.

Mostly at the moment it just validates that variables have been defined & that the types of everything make sense. I want to use it to implement type inference .which will be useful in the future for objects and such.